@@ -2,8 +2,8 @@
   "name": "@themost/node",
   "version": "2.5.1",
   "description": "MOST Web Framework NodeJS Client Module",
-  "module": "dist/themost_node.esm.js",
-  "main": "dist/themost_node.cjs.js",
+  "module": "dist/index.esm.js",
+  "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "test": "jasmine",
@@ -13,11 +13,11 @@
     "tslint": "tslint --project ."
   },
   "peerDependencies": {
-    "@themost/xml": "^2.5.1",
-    "@themost/client": "^2.5.6"
+    "@themost/client": "^2.5.6",
+    "@themost/xml": "^2.5.1"
   },
   "engines": {
-    "node": ">=6.9.4"
+    "node": ">=12.22.12"
   },
   "repository": {
     "type": "git",
@@ -30,17 +30,18 @@
   },
   "homepage": "https://github.com/themost-framework/node#readme",
   "dependencies": {
-    "node-fetch": "^2.6.0"
+    "node-fetch": "^2.7.0"
   },
   "devDependencies": {
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@themost/test": "^2.3.4",
     "@types/jasmine": "^3.5.0",
     "@types/node": "^12.12.14",
     "@types/node-fetch": "^2.5.4",
     "jasmine": "^3.10.0",
     "jasmine-spec-reporter": "^7.0.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.58.0",
-    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup": "^2.79.0",
     "ts-node": "^10.4.0",
     "tslint": "^6.1.3",
     "typescript": "^4.4.4"
