@@ -2,15 +2,18 @@ import {NodeDataContext, NodeDataService} from '../src';
 import {EdmSchema} from '@themost/client';
 import {getApplication, serveApplication, getServerAddress, getToken} from '@themost/test';
 import {URL, URLSearchParams} from 'url';
+import { Router } from 'express';
+import { Server } from 'http';
 
 const TEST_USER = 'alexis.rees@example.com';
 const TEST_PASSWORD = 'secret';
 
 describe('NodeDataContext', () => {
-    let liveServer;
-    let server_uri;
+    let liveServer: Server;
+    let server_uri: string;
+    let app: any;
     beforeAll(async () => {
-        const app = getApplication();
+        app = getApplication();
         liveServer = await serveApplication(app);
         server_uri = getServerAddress(liveServer);
     });
@@ -81,6 +84,36 @@ describe('NodeDataContext', () => {
         items.forEach( (item) => {
            expect(item.category).toBe('Laptops');
         });
+    });
+
+    it('should throw not found exception', async () => {
+        const context = new NodeDataContext( new URL('/api/', server_uri).toString(), {
+            useMediaTypeExtensions: false,
+            useResponseConversion: true
+        });
+        const token = await getToken(server_uri, TEST_USER, TEST_PASSWORD);
+        context.setBearerAuthorization(token.access_token);
+        await expectAsync(context.model('OtherProducts').getItems()).toBeRejectedWithError('Not Found');
+    });
+
+    it('should handle invalid response', async () => {
+        const token = await getToken(server_uri, TEST_USER, TEST_PASSWORD);
+
+        // add error router
+        const addRouter = Router();
+        addRouter.get('/api/Errors/', (req, res) => {
+            return res.status(409).send('<p>Conflict</p>');
+        });
+        // find first router
+        const route = app._router.stack.find((item: any) => item.name === 'router');
+        // and add sample error route at the beginning of its stack
+        route.handle.stack.unshift.apply(route.handle.stack, addRouter.stack);
+        const context = new NodeDataContext( new URL('/api/', server_uri).toString(), {
+            useMediaTypeExtensions: false,
+            useResponseConversion: true
+        });
+        context.setBearerAuthorization(token.access_token);
+        await expectAsync(context.model('Errors').getItems()).toBeRejectedWithError('Conflict'); //409
     });
 
 });
